Moved header files to include folder + added install rules
