+++ include/slam_karto/spa_solver.h
